+++ otx/algorithms/action/configs/classification/x3d/template.yaml
@@ -8,7 +8,7 @@ summary: Basic transfer learning template for X3D
 application: ~
 
 # Algo backend.
-framework: OTEDetection v2.9.1
+framework: OTXAction v2.9.1
 
 # Task implementations.
 entrypoints:
