+++ otx/algorithms/detection/adapters/mmdet/utils/config_utils.py
@@ -264,6 +264,7 @@ def patch_evaluation(config: Config):
     config.early_stop_metric = "mAP"
 
 
+# TODO Replace this with function in common
 def patch_data_pipeline(config: Config, template_file_path: str):
     """Update data_pipeline configs."""
     base_dir = os.path.abspath(os.path.dirname(template_file_path))
@@ -318,6 +319,7 @@ def cluster_anchors(config: Config, dataset: DatasetEntity, model: BaseDetector)
     return config, model
 
 
+# TODO Replace this with function in common
 @check_input_parameters_type()
 def get_data_cfg(config: Union[Config, ConfigDict], subset: str = "train") -> Config:
     """Return dataset configs."""
